Fixed dimension-mismatch bug for >2 visits
Fixed a dimension-mismatch bug in bayesweight() for > 2 visits. The bug arises as a non-conformable argument error during matrix multiplication. Previously, the index of only the previous treatment was included in beta_indices_s; the correct way is to include all prior-visit "bs" coefficients when nvisit > 1. Also corrected all_parameters to reflect correct marginal models when nvisit > 2.